langchain-aws/pyproject.toml
[project]
name = "spyglass-example-langchain-aws"
version = "0.1.0"
description = "LangChain AWS Bedrock example with Spyglass tracing and MCP tools"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-aws>=0.1.0",
    "langchain-core>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "python-dotenv>=1.0.0",
    "spyglass-ai",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Uncomment the section below to use the local spyglass-sdk for development
# This assumes spyglass-sdk is located at ../../spyglass-sdk relative to this file
# [tool.uv.sources]
# spyglass-ai = { path = "../../spyglass-sdk", editable = true }

